requirements.txt
requests==2.30.0
speedtest-cli==2.1.3
psutil==5.9.5
